requirements.txt
transformers
datasets
scikit-learn
pandas
torch
numpyflask
flask-cors
